// Client/src/app/pages/apps/debt/debt-graph/debt-graph.model.ts
// Chart data
export interface ChartType {
    chart?: any;
    plotOptions?: any;
    colors?: any;
    series?: any;
    stroke?: any;
    fill?: any;
    labels?: any;
    markers?: any;
    legend?: any;
    xaxis?: any;
    yaxis?: any;
    tooltip?: any;
    grid?: any;
    type?: any;
    sparkline?: any;
    dataLabels?: any;
    height?: any;
    option?: any;
    piechartcolor?: any;
    toolbar?: any;
    stacked?: any;
    color?: any;
    width?: any;
    padding?: any;
    fixed?: any;
    x?: any;
    y?: any;
    marker?: any;
    show?: any;
    curve?: any;
    zoom?: any;
    responsive?: any;
}